requirements.txt
Django==4.2.6
python-decouple==3.8
psycopg2==2.9.9
django-mptt==0.14.0